Move reducers up one level (#198)
